+++ packages/react-art/npm/lib/Rectangle.art.js
