+++ packages/react-art/npm/lib/Wedge.art.js
