+++ packages/react-art/src/ReactARTFiberEntry.js
